refactor: userAction & userReducer.js 이동
